requirements.txt
fastapi
pytest
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
cryptography
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
pydantic
pydantic-settings
redis>=4.0
slowapi
httpx